refactor to tree class
