+++ src/command/install.mjs
@@ -2,25 +2,21 @@
 import { Command } from 'commander'
 /* eslint-enable */
 import { requireRootProject } from '../project/project.mjs'
-import { reconcile, resolve } from '../project/dependencies.mjs'
 import { logger } from '../log.mjs'
+import { DependencyTree } from '../dependencies/dependency.tree.mjs'
 
 async function install () {
   const project = await requireRootProject()
 
   logger.progress('Resolving dependencies...')
-  const dependencies = await resolve(
+  const dependencyTree = await DependencyTree.resolve(
     project,
     (_, visited, remains) => logger.progress('Resolving dependencies...', visited / (visited + remains))
   )
-  logger.info('Resolved', dependencies.length, 'addons')
-  console.log(
-    dependencies.map(dep => [...dep.path, dep.source])
-      .map(path => '\t' + path.map(addon => `${addon.name}@${addon.version}`).join(' -> '))
-      .join('\n')
-  )
 
-  const depbag = reconcile(dependencies)
+  logger.info('Resolved', dependencyTree.dependencies.length, 'addons')
+
+  const depbag = dependencyTree.flatten()
   logger.info('Reconciled dependencies:\n', depbag.map(source => '\t' + source.stringify()).join('\n'))
 }
